// zeus-api/src/app/modules/designer/interfaces/shape-properties/shape-properties.interface.ts
export interface ShapeProperties {
  height: number;
  width: number;
  backgroundColorEnabled: boolean;
  backgroundColor: string;
  borderEnabled: boolean;
  borderColor: string;
  borderWidth: number;
  shadowEnabled: boolean;
  shadowColor: string;
  shadowX: number;
  shadowY: number;
  shadowBlur: number;
  opacity: number;
  visible: boolean;
}